+++ spec/e2e_spec.rb
@@ -29,7 +29,7 @@ describe 'E2E basic user story' do
     @s = Locatine::Search.new learn: true, visual_search: true, browser: Watir::Browser.new
     @s.browser.window.resize_to(1200, 1000)
     @s.browser.goto @path1
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.find(name: "element").text).to be == "Element"
     # Guess test
     expect(@s.find(name: "important span", vars: {"tag":"span", "text":"kadabra", "var":"che"}).text).to be == "Abrakadabra"
@@ -45,8 +45,8 @@ describe 'E2E basic user story' do
     @s.learn = true
     @s.browser = Watir::Browser.new
     @s.browser.goto @path6
-    expect(@s.collect(name: "lis fox").length).to be == 3
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.collect(name: "dogs")[0].text).to be == "ONE"
   end
 
@@ -56,7 +56,7 @@ describe 'E2E basic user story' do
 
   it "Finding elements" do
     @s.browser.goto @path2
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.find(name: "element", look_in: :labels).text).to be == "Element"
     iframe = @s.find(name: "iframe", look_in: :iframe)
     expect(@s.find(name: "h3 banana", iframe: iframe).text).to be == "Banana"
@@ -74,8 +74,8 @@ describe 'E2E basic user story' do
 
   it "can return a locator" do
     @s.browser.goto @path2
-    expect(@s.browser.elements(@s.lctr(name: "lis fox")).length).to be == 3
-    expect(@s.browser.elements(@s.lctr("lis fox")).length).to be == 3
+    expect(@s.browser.elements(@s.lctr(name: "lisa foxa")).length).to be == 3
+    expect(@s.browser.elements(@s.lctr("lisa foxa")).length).to be == 3
     expect(@s.browser.element(@s.lctr(name: "element")).text).to be == "Element"
     # Guess test
     expect(@s.browser.element(@s.lctr(name: "important span", vars: {"tag":"spoon", "text":"kedavra", "var":"ba"})).text).to be == "Abrakedavra"
@@ -88,28 +88,28 @@ describe 'E2E basic user story' do
 
   it "Finding exacts" do
     @s.browser.goto @path2
-    expect(@s.collect(name: "lis fox", exact: true).length).to be == 3
+    expect(@s.collect(name: "lisa foxa", exact: true).length).to be == 3
     expect(@s.find(name: "element", exact: true).text).to be == "Element"
   end
 
   it "Fails on exact when elements are lost" do
     @s.browser.goto @path3
-    expect(@s.collect(name: "lis fox", exact: true, no_fail: true, locator: {id: "not welcome"})).to be == []
-    expect(@s.check_collection(name: "lis fox", exact: true, no_fail: true)).to be == []
-    expect(@s.exact_collection(name: "lis fox", no_fail: true)).to be == []
+    expect(@s.collect(name: "lisa foxa", exact: true, no_fail: true, locator: {id: "not welcome"})).to be == []
+    expect(@s.check_collection(name: "lisa foxa", exact: true, no_fail: true)).to be == []
+    expect(@s.exact_collection(name: "lisa foxa", no_fail: true)).to be == []
     expect(@s.exact(name: "element", no_fail: true)).to be == nil
     expect(@s.find(name: "span for guess", no_fail: true, exact: true)).to be == nil
   end
 
   it "Fails when elements are lost and there is nothing similar" do
     @s.browser.goto @path4
-    expect{@s.collect(name: "lis fox")}.to raise_error(RuntimeError, 'Locatine cannot find element lis fox in Default')
+    expect{@s.collect(name: "lisa foxa")}.to raise_error(RuntimeError, 'Locatine cannot find element lisa foxa in Default')
     expect{@s.find(name: "element")}.to raise_error(RuntimeError, 'Locatine cannot find element similar to element in Default')
   end
 
   it "Will not fail with no_fail == true" do
     @s.browser.goto @path4
-    expect(@s.collect(name: "lis fox", no_fail: true)).to be == []
+    expect(@s.collect(name: "lisa foxa", no_fail: true)).to be == []
     expect(@s.find(name: "element", no_fail: true)).to be == nil
   end
 
@@ -122,7 +122,7 @@ describe 'E2E basic user story' do
     @s.browser.goto @path3
     @s.browser.window.resize_to(1200, 1000)
     expect(@s.collect(name: "strange collection sub").length).to be == 2
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.find(name: "element").text).to be == "Element"
     @s.visual_search = true
     expect(@s.find("h4 css").text).to be == "found anyway"
@@ -132,14 +132,14 @@ describe 'E2E basic user story' do
     a = @s.find("by coordinates", vars: {"x":"185", "y":"185"})
     expect(a.attribute('class')).to be == "korove"
     @s.browser.goto @path5
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
   end
 
   it "Ignoring unstable attributes" do
     Watir.default_timeout = 15
     start = Time.now
     @s.browser.goto @path2
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.find(name: "element").text).to be == "Element"
     expect(Time.now-start).to be < 10
   end
@@ -147,7 +147,7 @@ describe 'E2E basic user story' do
   it "Finds element if nesting structure is broken" do
     Watir.default_timeout = 3
     @s.browser.goto @path5
-    expect(@s.collect(name: "lis fox").length).to be == 3
+    expect(@s.collect(name: "lisa foxa").length).to be == 3
     expect(@s.find(name: "element").text).to be == "Element"
   end